Add azure default config
