--- gcp/logback-logstash-encoder-gcp/src/main/java/no/entur/logging/cloud/gcp/logback/logstash/StackdriverTimestampJsonProvider.java
@@ -1,30 +1,0 @@
-package no.entur.logging.cloud.gcp.logback.logstash;
-
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import com.fasterxml.jackson.core.JsonGenerator;
-import net.logstash.logback.composite.AbstractJsonProvider;
-import net.logstash.logback.composite.JsonWritingUtils;
-
-import java.io.IOException;
-
-/**
- * Add machine-readable timestamp.
- *
- * According to https://cloud.google.com/logging/docs/agent/logging/configuration#timestamp-processing:
- *
- * If the JSON timestamp representations is present in a structured record,
- * the Logging agent removes the fields from jsonPayload and collapses them into a single representation in the timestamp field in the LogEntry object.
- *
- */
-
-public class StackdriverTimestampJsonProvider extends AbstractJsonProvider<ILoggingEvent> {
-
-	@Override
-	public void writeTo(JsonGenerator generator, ILoggingEvent event) throws IOException {
-		generator.writeObjectFieldStart("timestamp");
-		generator.writeNumberField("seconds", event.getTimeStamp() / 1000);
-		generator.writeNumberField("nanos", event.getNanoseconds());
-		generator.writeEndObject();
-	}
-
-}
