+++ gcp/logback-logstash-encoder-gcp/src/main/resources/logback/logback-default-server.xml
@@ -1,13 +1,40 @@
 <included>
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverLogstashEncoder">
-            <fieldNames>
-                <version>[ignore]</version>
-                <timestamp>[ignore]</timestamp>
-            </fieldNames>
-            <provider class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverServiceContextJsonProvider">
-            </provider>
-            <provider class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverTimestampJsonProvider"/>
+        <encoder class="no.entur.logging.cloud.gcp.logback.logstash.AzureLogstashEncoder">
+            <providers>
+                <timestamp>
+                    <fieldName>timestamp</fieldName>
+                    <timeZone>UTC</timeZone>
+                </timestamp>
+                <loggerName>
+                    <fieldName>logger</fieldName>
+                </loggerName>
+                <logLevel>
+                    <fieldName>level</fieldName>
+                </logLevel>
+                <threadName>
+                    <fieldName>thread</fieldName>
+                </threadName>
+                <nestedField>
+                    <fieldName>mdc</fieldName>
+                    <providers>
+                        <mdc />
+                    </providers>
+                </nestedField>
+                <stackTrace>
+                    <fieldName>stackTrace</fieldName>
+                    <!-- maxLength - limit the length of the stack trace -->
+                    <throwableConverter class="net.logstash.logback.stacktrace.ShortenedThrowableConverter">
+                        <maxDepthPerThrowable>200</maxDepthPerThrowable>
+                        <maxLength>14000</maxLength>
+                        <rootCauseFirst>true</rootCauseFirst>
+                    </throwableConverter>
+                </stackTrace>
+                <message />
+                <throwableClassName>
+                    <fieldName>exceptionClass</fieldName>
+                </throwableClassName>
+            </providers>
         </encoder>
     </appender>
 </included>
