+++ gcp/spring-boot-autoconfigure-gcp-test/src/main/resources/logback/logback-spring-test.xml
@@ -10,12 +10,48 @@
 			<pattern>${CONSOLE_LOG_PATTERN}</pattern>
 			<charset>${CONSOLE_LOG_CHARSET}</charset>
 		</humanReadablePlainEncoder>
-		<humanReadableJsonEncoder class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverLogstashEncoder">
+		<humanReadableJsonEncoder class="no.entur.logging.cloud.gcp.logback.logstash.AzureLogstashEncoder">
 			<!-- remove unnecessary fields in testing -->
 			<!-- note: severity and severity is removed in code -->
 			<fieldNames>
 				<version>[ignore]</version>
 			</fieldNames>
+
+			<providers>
+				<timestamp>
+					<fieldName>timestamp</fieldName>
+					<timeZone>UTC</timeZone>
+				</timestamp>
+				<loggerName>
+					<fieldName>logger</fieldName>
+				</loggerName>
+				<logLevel>
+					<fieldName>level</fieldName>
+				</logLevel>
+				<threadName>
+					<fieldName>thread</fieldName>
+				</threadName>
+				<nestedField>
+					<fieldName>mdc</fieldName>
+					<providers>
+						<mdc />
+					</providers>
+				</nestedField>
+				<stackTrace>
+					<fieldName>stackTrace</fieldName>
+					<!-- maxLength - limit the length of the stack trace -->
+					<throwableConverter class="net.logstash.logback.stacktrace.ShortenedThrowableConverter">
+						<maxDepthPerThrowable>200</maxDepthPerThrowable>
+						<maxLength>14000</maxLength>
+						<rootCauseFirst>true</rootCauseFirst>
+					</throwableConverter>
+				</stackTrace>
+				<message />
+				<throwableClassName>
+					<fieldName>exceptionClass</fieldName>
+				</throwableClassName>
+			</providers>
+
 			<!-- add pretty-printing and coloring for testing -->
 			<!--
 				TODO this seems to add an extract space after the last newline,
@@ -31,14 +67,47 @@
 				</syntaxHighlighterFactory>
 			</jsonGeneratorDecorator>
 		</humanReadableJsonEncoder>
-		<machineReadableJsonEncoder class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverLogstashEncoder">
+		<machineReadableJsonEncoder class="no.entur.logging.cloud.gcp.logback.logstash.AzureLogstashEncoder">
 			<!-- note: this encoder will be used to evaluate assertions against log statements -->
 			<fieldNames>
 				<version>[ignore]</version>
 				<timestamp>[ignore]</timestamp>
 			</fieldNames>
-			<provider class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverServiceContextJsonProvider"/>
-			<provider class="no.entur.logging.cloud.gcp.logback.logstash.StackdriverTimestampJsonProvider"/>
+
+			<providers>
+				<timestamp>
+					<fieldName>timestamp</fieldName>
+					<timeZone>UTC</timeZone>
+				</timestamp>
+				<loggerName>
+					<fieldName>logger</fieldName>
+				</loggerName>
+				<logLevel>
+					<fieldName>level</fieldName>
+				</logLevel>
+				<threadName>
+					<fieldName>thread</fieldName>
+				</threadName>
+				<nestedField>
+					<fieldName>mdc</fieldName>
+					<providers>
+						<mdc />
+					</providers>
+				</nestedField>
+				<stackTrace>
+					<fieldName>stackTrace</fieldName>
+					<!-- maxLength - limit the length of the stack trace -->
+					<throwableConverter class="net.logstash.logback.stacktrace.ShortenedThrowableConverter">
+						<maxDepthPerThrowable>200</maxDepthPerThrowable>
+						<maxLength>14000</maxLength>
+						<rootCauseFirst>true</rootCauseFirst>
+					</throwableConverter>
+				</stackTrace>
+				<message />
+				<throwableClassName>
+					<fieldName>exceptionClass</fieldName>
+				</throwableClassName>
+			</providers>
 		</machineReadableJsonEncoder>
 	</appender>
